requirements.txt
PyYAML==6.0.1
Jinja2==3.1.2
web3==6.11.1
eth-account==0.9.0
eth-utils==2.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
click==8.1.7
rich==13.7.0
tabulate==0.9.0 